Add framework for mime decrypt / verify
Add event handling for Application and MailItem
events and use the message incoming handler in
the before read event.

* src/comhelp.h: Removed again. Macros moved to mymapi.h where
  they are better placed.
* src/eventsink.h: (EVENT_SINK_DEFAULT_DTOR_CODE): Release /
  unadvise properly.
  (USE_INVOKE_ARGS): New. Used to avoid unused parameter
  warnings.
  (debug_oom, debug_oom_extra): Moved to util.h.
* src/eventsinks.h: New. Header to declare event sink install /
  detach functions.
* src/eventsinks.cpp: New. Event sinks for MailItem and Application.
* src/oomhelp.h: Define GUID's necessary for the Event handlers.
  (get_object_by_id): New. oomhelp API style wrapper around
  QueryInterface.
* src/oomhelp.cpp (get_object_by_id): New.
* src/util.h (log_oom, log_oom_extra): New. Logging helper.
* src/mymapi.h: Add Macros from comhelp.
  (IMAPISecureMessage): Declare COM interface.
* src/message.h: Add headers that were indirectly included
  in other usages.
* src/gpgoladdin.h (GpgolAddin): Add Application Event sink member.
* src/gpgoladdin.cpp (OnStartupComplete): Install Application Event
  sink.

--

This establishes the framework for proper MIME handling
currently mails are decrypted / verified but the message
content is not yet shown in Outlook.

The ApplicationEvent sink handles ItemLoad events (which should
happen once a Message is loaded into Outlook) and uses the event
to connect a MailItemEvent handler to the MailItem. The
MailItemEvent handler then handles the BeforeRead event and