Add support for MIME encryption / signing
This currently tries to encrypt / sign all mails sent.

* src/eventsinks.cpp (MailItemEvents::Invoke): Sign/Encrypt
  mails in AfterWrite event.

--
Before this I tried using the Application's ItemSend event and
the Send / Write event's of the MailItem. In the ItemSend
event you can't access the SecureItem interface and the
MAPIObject is not filled.

In Send / Write get_oom_base_message works but again
the relevant values are not set. Only AfterWrite gives
us access to the filled MAPI Object but in case of errors
we don't have a cancel parameter here anymore to
abort sending the mail. So this might not be the best
event for this.