Hopefully fix Ol 2016 crash on send
* src/mailitem-events.cpp (request_send): Remove delay.

--
Earlier tests with 2013 made me think that the delay was necessary
here. But Ol 2016 and Ol 2013 work reliably without it.
This makes the code more deterministic (Placing a Window message
in the Queue right after the encryption is done.) and appears
to fix a random crash in Ol 2016 where Outlook Itself would crash
after sending the message.