Fix body handling of sent mails.
* src/mailitem-events.cpp (handle_read): Work on the base message.

--
Outlook somehow is confused about the attachment
table of our sent mails. The securemessage interface
gives us access to the real attach table but the attachment
table of the message itself is broken.

So unprotect attachments does not work for sent mails
as the attachment table of the mapiitem is invalid.
We need to somehow get outlook to use the attachment table
of the base message and and then decrypt those.
This will probably mean removing all attachments for the
message and adding the attachments from the base message then
we can call unprotect_attachments as usual.

For now this works only for the body.