Really fix msg merge failures.
* src/Makefile.am: Do not link dialog.o into gpgol.dll. It's
 already included in resource.o.
* src/dialogs.h: Remove comment about not using the lower 4 bit.
* src/oomhelp.cpp (put_oom_icon): Update accordingly. (To use +1 again)

--
This fixes the problem 5d62fdbf was supposed to fix.
The "28 bit" ID part was just a red herring. The whole merge
code was triggered because we linked the resources twice into
the dll. Once through dialog.o and once through resource.o

The fix from 5d62fdbf only worked because changes to the id did
not cause the dialog object to be rebuild and so the problem
disappeared for me because dialog.o and resource.o used different
ID's. A clean build triggered the problem again.