Use UTF8 in addin localization.
* src/w32-gettext.c, src/w32-gettext.h (utf8_gettext): New.
  gettext that returns UTF8.
* src/w32-gettext.c (internal_gettext): Helper that can return
  either UTF8 or native encoding.
* src/gpgoladdin.cpp: Use UTF8 gettext variants.

--
Using native encoding is always problematic on Windows and should
be avoided. Native encoding can't represent Unicode and conversions
to wchar are bound to lose data. This usually breaks in a lot of ways
with Unicode filenames etc.

I like to use UTF8 internally for all strings and convert them to
UTF 16 (wchar) for Windows API use.

For compatibility reasons (not to break the old Exchange Extension)
the default gettext still returns native encoding and you have to
redefine _ to get the utf8 gettext.