More logging and error checking in App Events
* src/application-events.cpp (ApplicationEvents::Invoke): Check
 that parms is not NULL. Add more debug output.

--
According to the Windows Event log attached to gnupg-bug: 1837
gpgol caused a crash while handling 0xFBA7 (ItemLoad).
This should help to home in on such a crash. I doub't that
parms is NULL in the Invoke call but as you can never
know let's be more defensive with error checks.