Make old decrypt / verify buttons work with mime
* src/ribbon-callbacks.cpp (verify_mime): New. Force message handling
 again.
 (do_reader_action): Call verify_mime on action and print info
 on decrypt.

--
I've tried to set the body to the plain text again and decrypt
the attachments in that case too. But It's no good. There are
several issues with it. We can't reliably detect if we need
to wipe the message as outlook sometimes does not give us the
MAPI Object in the write event. And Attachments are not properly
updated after decryption.
The Read event still appears the only place for us to correctly
insert the plain text.