Fix handling sent and exchange to exchange mails.
* src/mail.cpp (Mail::process_message): Always change message class.

--
Changing the message class in the before read event causes
Outlook to reconsider what it knows about a message and to read
information again from the unwrapped basemessage. While this
may have a negative impact on performance this fixes the case
where Outlook already knows too much about a message because
it was either recieved by exchange (and so already had the
MAPI properties from GPGOL set) or because it sent the
message itself and so also had the GpgOL MAPI properties.