Add rfc2047 parser based on GMime
* src/rfc2047parse.c, src/rfc2047parse.h: New.
* src/Makefile.am: Update accordingly.
--
To properly handle attachment names the same way as outlook
does we need to be able to parse rfc2047 words at least.
From the rfc's I'm not sure if we need to parse multiple words
but let's assume we do for added compatibility. And this was
an easy part of GMime to extract.

For now it only handles UTF8 binary and quoted printable.