Start UIServer on demand
* src/engine-assuan.c (op_assuan_init): Don't test connection.
--
The old spawning code with the retry counter made this unfeasible.
Starting on demand has the charm that it is properly handled
if the user stops the UIServer or the UIServer crashes (which of course
only Kleo does).

This should also give us a big boost in the startup timing benchmark
Outlook runs to suggest "Addins that are making Outlook slow".

GnuPG-Bug-Id: 2137