Add mail revert implementation that works on OOM
* src/revert.cpp (gpgol_mailitem_revert): Revert mail in OOM.
* src/revert.h: Add prototype.

--
Using OOM appeared to be necessary. At least I have not found
a way to modify the Body through MAPI in a way that it survived
the write event. It appeared that during the Write event the
OOM Body is synced back to mapi. Same for the attachment model
in OOM. Some weirdeness there. Now we use OOM as much as possible.

There is some duplication with the old code but I didn't want
to touch the old code to aviod regressions. Although I doubt
very much that the old code works correctly.

For now the code returns an error for all messages apart
from PGP Message.