Try harder to save reverted mails
* src/revert.cpp (finalize_mapi): Try several secons to store.

--
Sometimes we get an error object has changed here. This
appears to be related to other MAPI io happening, and coincides
with a busy disk. So we do some evil sleeping in the GUI thread.
This actually works. If we don't and have errors
there we are in a bad place as we don't know if the message has
been reverted.