Add a way to resolve Exchange Address in MAPI
* src/mapihelp.cpp (resolve_ex_from_address): New.
 (mapi_get_from_address): Use it.

--
This fixes the case for decrypt where we can't access the
Mailitem to get the sender because we decrypt / verify
before the Mailitem is loaded.
The PidTagSenderSmtpAddress_W and
PR_SENT_REPRESENTING_SMTP_ADDRESS_W properties used before
do not work for mails in the sent mails folder,.. There
was no property from which the SMTP Adress could be extracted.