Try to get exchange address from cache first.
* src/mapihelp.cpp (resolve_ex_from_address): Try cache first.
* src/mymapi.h (MAPI_CACHE_ONLY): Define.

--
According to MSDN there is a value for MAPI_CACHE_OK which would
basically do what we want. Look in the cache and if it's not cached
look it up from the server. But the value is not documented
and also not in OutlookSpy.

This fixes a hang when the exchange connection is not good.