Start refactoring mimeparser into mimedataprovider
* src/mail.cpp (Mail::parsing_done): Use new API.
* src/mailparser.cpp: Move out body / attachment handling.
* src/mailparser.h: Update accordingly.
* src/mapihelp.cpp (mapi_mark_or_create_moss_attach): Fix return
for multipart/signed.
* src/mimedataprovider.cpp, src/mimedataprovider.h: Rework.

--
The MimeDataprovider now handles the transition from a MIME Message
into a signed / signature or for crypto messages into a (html)body
with attachments. It can be wrapped into GpgME::Data.

After creation of a MimeDataprovider no more MAPI structures
are needed and everything is currently done in memory.

The mimeparser code is based on the old mimeparser.c but a bit
simplyfied and restructured to split out the actual mime parsing
from the crypto calls.

This is a first working version and is not fully working
yet.