Remove AfterWrite Windowmessage hack
* src/mailitem-events.cpp (EVENT_SINK_INVOKE): Call save in
the send event to trigger the write and then send the mail.
(request_send): Removed.
* src/Windowmessages.cpp (request_send_mail): Removed.
src/Windowmessages.h (REQUEST_SEND_MAIL): Removed.

--
The old code was written under the bad assumption that
after write we could not send the mail directly.

This should fix the reported crashes when sending mails:

GnuPG-Bug-Id: 2335