Add explorer and explorers event handler
* explorer-events.cpp: New. Invalidate ui on selection change.
* explorers-events.cpp: New. Install event handler for new explorers.
* gpgoladdin.cpp (GpgolAddin::~GpgolAddin): Release explorers event
handler.
(install_explorer_sinks): Install explorers and explorer event
handler.
(GpgolAddin::OnStartupComplete): call install_explorer_sinks

--
This fixes the problem that without another read by outlook
the state of the ui might not have matched to the signature
state, which is critical as it may have shown the wrong sigstate.