Only close on WM_CLOSE of explorers
* src/explorer-events.cpp (EVENT_SINK_INVOKE): Track explorers.
* src/gpgoladdin.cpp (install_explorer_sinks): Install sinks.
* src/windowmessages.cpp (add_explorer_window)
(remove_explorer_window): New.
(gpgol_hook): Only react to explorers.

--
This prevents mails from beeing closed when they are opened
in a reader. Still todo is to close only the mails from the
selection of the explorer that is currently closing and
only to close all mails if it is the last explorer.