Start locate-keys when a recipient is added
* src/mail.cpp (Mail::locate_keys): Check for recipient keys.
(do_locate): New.
(Mail::needs_crypto): Return flags.
* src/mail.h: Update accordingly.
* src/mailitem-events.cpp (EVENT_SINK_INVOKE): Handle recipient
changes and do locate_keys if necessary.

--
Checked addresses are stored in a set so that they are not
searched twice per session to avoid spammy requests.

This feature is mostly useful for users of auto-key-locate
where this can be used to retrieve keys from dane etc.