Ignoore WM_SYSCOMMAND / SC_CLOSE
* src/windowmessages.cpp (gpgol_hook): Ignore WM_SYSCOMMAND
SC_CLOSE.

--
I have not found a way where SC_CLOSE is sent but WM_CLOSE
is not sent. This fixes double close crashes when both WM_SYCOMMAND
and SC_CLOSE are handled but the close did not complete the
unload of the message.

Maybe we should remove our mails from the internal map as soon
as we call close.