Don't close non-crypto mails
* src/mail.cpp (Mail::close_all_mails): Ignore non crypto mails.
Don't delete mail on error.

--
Closing non crypto mails is a chance for data loss as it could
have caused drafts to be discarded. The delete on error is also
not good as we don't know if the close already triggered the
unload event and so may have unloaded the mail already.