Allow different UUIDs for the same mail
* src/oomhelp.cpp (get_unique_id): Always create one if
called with create parameter.

--
Outlook can have multiple Event Handler for the Same Mail.
E.g. If you attach a mail to another mail it becomes so decoupled
that it has it's own event handler.

We must handle that case to avoid messing up or event handling /
internal state. It is now:
One UUID for One mailitem-events handler.