Warn if partial crypto with attachments is found
* src/mail.cpp (Mail::check_attachments): Check if
all attachments are hidden and warn otherwise.
(Mail::parsing_done): Call check_attachments.
* src/mail.h: Update accordingly.
* src/oomhelp.p: Add new DASL.

--
This is important for inline Crypto mails where Attachments
may not have been signed or encrypted.

For MIME Mails we don't have the problem as we only handle
Crypto MIME Mails if the top level mime part is a crypto part
and so everything is crypto or not.