Delete Body MAPI Props after MIME creation
* src/mimemaker.cpp (finalize_message): Delete Body Props.

--

Somehow this is now required. For outgoing mails over SMTP
the body is removed when the mime structure of our mime attachment
is used. But for exchange <> exchange or in the sent mail
folder this no longer happens. Causing a plain text leak.

We now explicitly delete the MAPI Body after our attachment
is created.

GnuPG-Bug-Id: T3369