Fix regression for multipart/signed mails
* src/mapihelp.cpp (mapi_change_message_class): Call
change_message_class_ipm_note_smime_multipartsigned even
if S/MIME is disabled.

--
This function actually checks if a message with the class
IPM.Note.SMIME.MultipartSigned is actually a PGP / MIME mail
so it needs to be called even if S/MIME is disabled.

This fixes multipart signed mail handling.