Fix reversion for S/MIME mails if S/MIME disabled
* src/mailitem-events.cpp (EVENT_SINK_INVOKE): Revert
again if S/MIME is disabled but we parsed the mail.

--
While reversion is no longer neccessary for crypto mails
to hide the plaintext we still might need it to change
the message class back. And wiping the body etc. is also
a good idea when we "leave a mail alone"

GnuPG-Bug-ID: T3494