Improve / Fix UI Invalidation
* src/explorers-events.cpp (EVENT_SINK_INVOKE): Don't cancel
invalidation when preview is disabled. Call delayed_invalidate_ui.
* src/mailitem-events.cpp (EVENT_SINK_INVOKE): Use
delayed_invalidate_ui.
* src/windowmessages.cpp (delayed_invalidate_ui): New.

--
This fixes invalidation in that it is no longer prevented
if the reading pane is not found on selection change and by
using the delay with a guard for an already happening
invalidation it should fix loops / races of many invalidations.

GnuPG-Bug-Id: T3523
GnuPG-Bug-Id: T3535