Fix multiple selection performance
* src/ribbon-callbacks.cpp (get_mail_from_control): Treat
multiselection like no mail is selected.

--
Experiments with OutlookSpy and debug output have shown that
the call:
    mailitem = get_oom_object (context, "Selection.Item(1)");
triggered an ItemLoad event for each mail in a selection. This
is fun when there are thousands of mails selected and can
lead to out of resources. As a workaround for this Outlook
bug we now treat multi selection as if no mail was selected
and don't obtain the first item.

GnuPG-Bug-Id: T3523
GnuPG-Bug-Id: T3433