Be more lenient with broken PGP Inline messages
* src/mimedataprovider.cpp (MimeDataProvider::collect_data):
Fix up PGP Messages.

--
In the Outlook world sometimes messages are rewritten / converted
to HTML and then converted back to Plaintext etc.
This can lead to broken PGP Messages where there is extra
whitespace. We now fix up the whitespace and remove comments
in the Mimedataprovider in the spirit of trying to accept
as much input as possible.

GnuPG-Bug-Id: T3821