Stop prasing on non whitepspace before PGP Message
* src/mapihelp.cpp (get_msgcls_from_pgp_lines): Again break
on non whitespace before message marker.

--
The problem here is that for inline messages a full quote in
the response could cause GpgOL to use the full quote as the
message instead of the reply. Outlook itself formats
its quotes in a way that would cause this.

While this might break again MTA's which insert stuff above
a Message we have to do this.

GnuPG-Bug-Id: T3964