Handle always sign/encrypt for inline reply/fw
* src/mailitem-events.cpp (reply, forward, replyall): Handle
the always sign/encrypt options.

--
If the option reply_crypt is set and the reply is to
a crypto mail the settings from that mail will override
the always options.

Previously this only worked in the Open command so only
if a new message composer window was opened.

This fixes a problem noticed in:
GnuPG-Bug-Id: T4037