Improve handling of UID origin
* src/mail.cpp (Mail::getCryptoDetails_o, Mail::updateSigstate):
Treat WKD origin as "shortcut to level 2.

--
As described on https://wiki.gnupg.org/AutomatedEncryption we
treat WKD already as a shortcut to level 2 as we have an
HTTPS secure indication from the provider that the key matches
the mail address.

This sets the green bar "verified sender" already for such mails,
because the senders address was verified through the provider.

To protect against phishing attacks this does not help but for
these Level 3 and 4 are required.