Disable async encryption for mails with attachs
* src/mail.cpp (check_inline_response): Disable async crypt
if mail has attachments.
* src/windowmessage.cpp (CRYPTO_DONE): Clarify that the
fallback to MAPI is a very very last resort.

--
This commit sacrifices usability (async encryption) for
stability. We now use the old sync crypt code when attachments
are added to a crypto mail to avoid errors that occured
for some attachments.
As there does not appear to be a detectable pattern or
value for which attachments the write in the second send
fails we block async crypt for all attachments.

GnuPG-Bug-Id: T4131