Show signer KeyID and UserID for valid mails
* src/mail.cpp (Mail::updateSigFlag_o): Set the flag.
(Mail::flagChangeTriggered): New helper to carry the state.
(pretty_id): Helper to format a fingerprint / long keyid.
* src/mailitem-events.cpp (PropertyChange): Ignore our
own changes.

--
This has two purposes. If the name of the sender is large
or looks like a mail address it is not visible at first
glance which key was used to verify as the real mail
address might be out of the screen.

Additionally this serves as another split between the
Verified Sender category and the body to make it even
harder to undetectable fake this through HTML. It
is never possible to fake it or replace it together
with the changed status icon.