Pass write for closed mails
* src/folder-events.cpp (Invoke): Set is now implicit in close.
* src/mail.cpp: Change isAboutToBeMoved to passWrite.
(decryptVerify_o): Stop passing write when processing.
* src/mailitem-events.cpp (Write): Pass it when a mail was closed
by us.

--
After closing with discard changes it should be ok to write
as long as we are not decrypted again. This can be helpful
after an external close was triggered and is basically
the same as what we do for a move.

GnuPG-Bug-Id: T4241