Check for error in dbg decrated oom funcs
* src/oomhelp.cpp (gpgol_openProperty, gpgol_queryInterface):
Check result before calling memdbg_addRef.

--
Calling something on the pointer on error is undefined
so we may not do it. memdbg_addRef tries to get the
name of the interface. That might then crash.
Usually the returned pointer should be NULL so
on error so it does not happen much.

GnuPG-Bug-Id: T4262