Add minimalistic protected-headers support
* src/mail.cpp (Mail::parsing_done): Update subject if one
was taken from a protected header.
* src/mailitem-events.cpp (prop_blacklist): Ignore subject related
changes.
* src/mimedataprovider.cpp (t2body): Handle protected headers.
(MimeDataProvider::finalize): Convert subject to a string.
(MimeDataProvider::get_internal_subject): New.
* src/parsecontroller.cpp (ParseController::get_internal_subject):
New accessor.
* src/parsecontroller.h, src/mimedataprovider.h: Update accordingly.

--
This is still bad but a little less bad then it was before.
We still ignore to and from from the protected headers but at
least we show the subject now.1