Fix / improve decrypting sent exchange mails again
* src/mail.cpp (Mail::preProcessMessage_m): Also pass the
parsed_message.
* src/mapihelp.cpp (hide_attachment_mapipos): New helper.
(mapi_mark_or_create_moss_attach): Take parsed msg as
additional parameter. Improve MOSSTMPL handling.

--
If we have a mosstempl. This means that we created
this message. Sometimes we get the situation in
sent mails that we have a mosstempl but the
pgp version header and octet streams also
attached in the parsed_message. We can't
trigger a reread from the base message so
in that case we want to hide everything
as we rebuild the message from our MOSS.

We don't delete here to avoid bugs where data loss
might occur.

GnuPG-Bug-Id: T4241