Use IPM.Note.SMIME.MultipartSigned again
* src/mimemaker.cpp (finalize_message): Use standard form
for S/MIME.

--
When sending on the same exchange server the mapi properties
and so our message class is kept when sending. So
another (faulty) client without GpgOL might receive this message
with the InfoPathForm MessageClass and might not
handle it as an S/MIME mail.

According to MS-OXOSMIME IPM.Note.SMIME should also be possible
for encrypted mails.
For now we use MultipartSigned again.