Only autosecure when seckey found for CMS
* src/keycache.cpp (isMailResolvable): If the prefer smime
option is selected only autosecure if there is a secret CMS cert.

--
When OpenPGP keys could be found for all recipients
it would toggle autosecure even if S/MIME was preferred.
As we don't have a good way to automatically handle
SMIME key generation we should not automatically toggle
encryption for that case.

GnuPG-Bug-Id: T4483