Fix forwarding of plain text mails with CID attachs
* src/mimemaker.cpp (add_body_and_attachments): Handle non
related case better.
(write_attachments): Do not filter attachments with CID in
non related case.

--
Before this the code had the assumption that a content-id
could only exist on an attachment for HTML mails as it otherwise
does not make sense.

But when forwarding mails with attachments they somehow
get a content id and thus in the case of plain text mails
they were filtered out.

GnuPG-Bug-Id: T4526