Saveguard against related attachment hiding
* src/mail.cpp (fixup_last_attachment_o): Search for
cid in html body before setting it.

--
As outlook hides elements with a content id and
we cannot seem to prevent that we now make extra
sure that the content id is actually referenced
in the mail body before setting it.

This fixes an issue when forwarding attachments
and might also be the reason for other hidden
attachments.

GnuPG-Bug-Id: T4203
GnuPG-Bug-Id: T4526